test: remove low-signal helper coverage
